Working timer feature
A working timer feature using recursive singleShot QTimers. Buttons aside from Stop Button are disabled to prevent user interference while waiting.

To do: display a countdown or elapsed timer.
To do: get the Stop Button to interrupt the singleshot qtimer. As-is, the Stop Button should halt the pump but the timer continues in the background, and is caught when it attempts to repeat again.

#40